keychanin: Enhance logic and introduce axzsh_keychain_update function
Now keychain(1) is called in the "profile" stage for login shells
already, to make sure that it runs before any other local scripts
test, use, or update the SSH and GnuPG agents.

And in addition this patch introduces the new axzsh_keychain_update
function, which can be used to update the SSH and GnuPG agent
information at any time.