New "30_env" core plugin
Move setting of XDG_CACHE_HOME and ZSH_CACHE_DIR from 20_home to this
new plugin, and set XDG_RUNTIME_DIR, too.

In addition, make sure that TMPDIR is reset when it has been deleted but
TMP is available (this is the case for setgid binaries on Linux, for
example, which is quite common when using screen(1)).