Cache: Read "zprofile" in the correct context in "zshrc" stage
When in the "zshrc" stage, make sure that the "zprofile" stage has
already been handled. This uses the "01_zprofile" plugin which is used
in the "zshrc.cache" file itself as well, but can't be used successfully
there because it becomes sourced inside of a ZSH function; so we have to
source it in the global context manually ...

And use the cache "zprofile.cache" in the "01_zprofile" plugin, when
available.