std_functions: Enhance take(), introduce untake()
take() now creates an new temporary directory when called without
parameters (and changes the working directory into it).

The new untake() functions changes the current working directory to the
home directory and recursively deletes the old working directory when
the path matches a pattern which is used for temporary directorues
created by take().