11_terminal: "Downgrade" TERM when it does not match TERM_COLORS
When TERM indicates that it supports 256 colors but TERM_COLORS does
not, we "downgrade" TERM and save its original value in the new
TERM_DOWNGRADED_FROM environment variable.

Terminals on which TERM_DOWNGRADED_FROM is set are regaded as "not
modern" by the axzsh_is_modern_terminal() function, which in turn
disables some "fancy" plugins that only make sense (or work properly)
when 256 or more colors are available.

The "axttyinfo" command shows the TERM_DOWNGRADED_FROM value after the
terminal size in square brackets when set.