homebrew: Cache "brew shellenv" for 1 day
As of 2024-11-26, the brew command (e.g. /opt/homebrew/bin/brew) calls
"sudo --reset-timestamp" which is, in theory, a very good idea but
results in long delays if your machine is, for example, linked to Active
Directory and your domain controller is not available or slow (not sure
why sudo(1) actually checks any user credentials for _forgetting_ the
usage time, but that's how it seems to be). Anyway. This can lead to
annoying delays spawning new sessions, so use a "cache file" to speed
this up.