ssh_autoadd: Integrate functionality of "ssh_macos" into "ssh_autoadd"
This allows us to better use the two together: now on Darwin/macOS,
we first try to loads keys using passwords from the "keychain", and
it that fails (=no keys were loaded), the regular "ssh-autoadd" behavior
kicks in: we run "plain ssh-add".

And the latter now not only runs on a connected terminal, but also when
the DISPLAY or SSH_ASKPASS environment variables are set.